fix: Alarm doesn't ring when `Show Motivational Quote` is turned on. (#804)
* fixed: problem with showMotivationalQuote

* removed unnecessary variable

---------
