+++ lib/app/modules/alarmRing/controllers/alarm_ring_controller.dart
@@ -9,6 +9,7 @@ import 'package:flutter_volume_controller/flutter_volume_controller.dart';
 
 import 'package:get/get.dart';
 import 'package:ultimate_alarm_clock/app/data/models/alarm_model.dart';
+import 'package:ultimate_alarm_clock/app/data/models/quote_model.dart';
 
 import 'package:ultimate_alarm_clock/app/data/models/user_model.dart';
 
@@ -16,7 +17,9 @@ import 'package:ultimate_alarm_clock/app/data/providers/firestore_provider.dart'
 import 'package:ultimate_alarm_clock/app/data/providers/isar_provider.dart';
 import 'package:ultimate_alarm_clock/app/data/providers/secure_storage_provider.dart';
 import 'package:ultimate_alarm_clock/app/modules/settings/controllers/settings_controller.dart';
+import 'package:ultimate_alarm_clock/app/modules/settings/controllers/theme_controller.dart';
 import 'package:ultimate_alarm_clock/app/utils/audio_utils.dart';
+import 'package:ultimate_alarm_clock/app/utils/constants.dart';
 
 import 'package:ultimate_alarm_clock/app/utils/utils.dart';
 import 'package:vibration/vibration.dart';
@@ -36,6 +39,7 @@ class AlarmControlController extends GetxController {
   RxBool showButton = false.obs;
   StreamSubscription? _sensorSubscription;
   HomeController homeController = Get.find<HomeController>();
+  ThemeController themeController = Get.find<ThemeController>();
   SettingsController settingsController = Get.find<SettingsController>();
   RxBool get is24HourFormat => settingsController.is24HrsEnabled;
   Rx<AlarmModel> currentlyRingingAlarm = Utils.alarmModelInit.obs;
@@ -49,9 +53,6 @@ class AlarmControlController extends GetxController {
   late Timer guardianTimer;
   RxInt guardianCoundown = 120.obs;
   RxBool isPreviewMode = false.obs;
-
-
-
 
   getNextAlarm() async {
     UserModel? _userModel = await SecureStorageProvider().retrieveUserModel();
@@ -169,6 +170,68 @@ class AlarmControlController extends GetxController {
     });
   }
 
+  void showQuotePopup(Quote quote) {
+    Get.defaultDialog(
+      title: 'Motivational Quote',
+      titlePadding: const EdgeInsets.only(
+        top: 20,
+        bottom: 10,
+      ),
+      backgroundColor: themeController.secondaryBackgroundColor.value,
+      titleStyle: TextStyle(
+        color: themeController.primaryTextColor.value,
+      ),
+      contentPadding: const EdgeInsets.all(20),
+      content: Column(
+        children: [
+          Obx(
+            () => Text(
+              quote.getQuote(),
+              style: TextStyle(
+                color: themeController.primaryTextColor.value,
+              ),
+            ),
+          ),
+          const SizedBox(
+            height: 15,
+          ),
+          Align(
+            alignment: Alignment.centerRight,
+            child: Obx(
+              () => Text(
+                quote.getAuthor(),
+                style: TextStyle(
+                  color: themeController.primaryTextColor.value,
+                  fontWeight: FontWeight.w600,
+                  fontStyle: FontStyle.italic,
+                ),
+              ),
+            ),
+          ),
+          const SizedBox(
+            height: 30,
+          ),
+          TextButton(
+            style: ButtonStyle(
+              backgroundColor: MaterialStateProperty.all(
+                kprimaryColor,
+              ),
+            ),
+            onPressed: () {
+              Get.back();
+            },
+            child: Text(
+              'Dismiss',
+              style: TextStyle(
+                color: themeController.secondaryTextColor.value,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
   @override
   void onInit() async {
     super.onInit();
@@ -199,8 +262,6 @@ class AlarmControlController extends GetxController {
         }
       });
     }
-
-
 
     showButton.value = true;
     initialVolume = await FlutterVolumeController.getVolume(
@@ -268,6 +329,12 @@ class AlarmControlController extends GetxController {
     // }
 
     AudioUtils.playAlarm(alarmRecord: currentlyRingingAlarm.value);
+
+
+    if(currentlyRingingAlarm.value.showMotivationalQuote) {
+      Quote quote = Utils.getRandomQuote();
+      showQuotePopup(quote);
+    }
 
     // Setting snooze duration
     minutes.value = currentlyRingingAlarm.value.snoozeDuration;
