+++ lib/app/modules/home/controllers/home_controller.dart
@@ -274,14 +274,6 @@ class HomeController extends GetxController {
       scalingFactor.value = (minFactor + (maxFactor - minFactor) * newFactor);
     });
 
-    if (Get.arguments != null) {
-      bool showMotivationalQuote = Get.arguments.showMotivationalQuote;
-
-      if (showMotivationalQuote) {
-        Quote quote = Utils.getRandomQuote();
-        showQuotePopup(quote);
-      }
-    }
   }
 
   refreshUpcomingAlarms() async {
@@ -592,68 +584,6 @@ class HomeController extends GetxController {
         colorText: Colors.white,
       );
     }
-  }
-
-  void showQuotePopup(Quote quote) {
-    Get.defaultDialog(
-      title: 'Motivational Quote',
-      titlePadding: const EdgeInsets.only(
-        top: 20,
-        bottom: 10,
-      ),
-      backgroundColor: themeController.secondaryBackgroundColor.value,
-      titleStyle: TextStyle(
-        color: themeController.primaryTextColor.value,
-      ),
-      contentPadding: const EdgeInsets.all(20),
-      content: Column(
-        children: [
-          Obx(
-            () => Text(
-              quote.getQuote(),
-              style: TextStyle(
-                color: themeController.primaryTextColor.value,
-              ),
-            ),
-          ),
-          const SizedBox(
-            height: 15,
-          ),
-          Align(
-            alignment: Alignment.centerRight,
-            child: Obx(
-              () => Text(
-                quote.getAuthor(),
-                style: TextStyle(
-                  color: themeController.primaryTextColor.value,
-                  fontWeight: FontWeight.w600,
-                  fontStyle: FontStyle.italic,
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(
-            height: 30,
-          ),
-          TextButton(
-            style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.all(
-                kprimaryColor,
-              ),
-            ),
-            onPressed: () {
-              Get.back();
-            },
-            child: Text(
-              'Dismiss',
-              style: TextStyle(
-                color: themeController.secondaryTextColor.value,
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
   }
 
   Future<void> swipeToDeleteAlarm(UserModel? user, AlarmModel alarm) async {
